memento-agent: extract retrieveContext from MementoAgent
