+++ packages/memento-agent/src/mementoAgent.ts
@@ -1,28 +1,20 @@
 // Path: packages/memento-agent/src/mementoAgent.ts
+
 import {  type SendMessageArgs } from "@memento-ai/conversation";
-import { functionCallingInstructions } from "./dynamicPrompt";
 import { Agent, type AgentArgs, type SendArgs } from "@memento-ai/agent";
-import { type FunctionCallResult } from "@memento-ai/function-calling";
+import { awaitAsyncAgentActions, startAsyncAgentActions } from "./asyncAgentGlue";
 import { getDatabaseSchema } from "@memento-ai/postgres-db";
+import { invokeSyncAndAsyncFunctions, type InvokeFunctionsResults, FUNCTION_RESULT_HEADER, type InvokeFunctionsArgs } from "./invokeFunctions";
+import { mementoPromptTemplate, type MementoPromptTemplateArgs } from "./mementoPromptTemplate";
 import { registry } from "@memento-ai/function-calling";
+import { retrieveContext } from "./retrieveContext";
+import { type FunctionCallResult } from "@memento-ai/function-calling";
 import { USER, type Message, type Role, type Memento } from "@memento-ai/types";
 import { Writable } from "node:stream";
 import c from 'ansi-colors';
-import debug from "debug";
-import Handlebars from "handlebars";
-import type { Context, SimilarityResult } from "@memento-ai/memento-db";
-import type { SynopsisAgent } from "@memento-ai/synopsis-agent";
-import { mementoCoreSystemPrompt } from "./mementoCoreSystemPrompt";
-import { mementoPromptTemplate, type MementoPromptTemplateArgs } from "./mementoPromptTemplate";
-import { invokeSyncAndAsyncFunctions, type InvokeFunctionsResults, FUNCTION_RESULT_HEADER, type InvokeFunctionsArgs } from "./invokeFunctions";
-import { awaitAsyncAgentActions, startAsyncAgentActions } from "./asyncAgentGlue";
+import type { Context } from "@memento-ai/memento-db";
 import type { ContinuityAgent } from "@memento-ai/continuity-agent";
-
-Handlebars.registerHelper('obj', function(context) {
-    return Bun.inspect(context);
-});
-
-const mlog = debug("mementoAgent:mem");
+import type { SynopsisAgent } from "@memento-ai/synopsis-agent";
 
 export type MementoAgentArgs = AgentArgs & {
     outStream?: Writable;
@@ -37,18 +29,18 @@ export type MementoAgentArgs = AgentArgs & {
 
 export class MementoAgent extends Agent
 {
-    private databaseSchema: string;
-    private outStream?: Writable;
-    private continuityAgent?: ContinuityAgent;
-    private synopsisAgent?: SynopsisAgent;
-    private lastUserMessage: string;
-    private max_message_pairs: number;
-    private max_csum_tokens: number;
-    private max_similarity_tokens: number;
-    private max_synopses_tokens: number;
-    private asyncResults: Promise<FunctionCallResult[]>;
-    private continuityResponsePromise: Promise<string> | null = null;
-    private continuityResponseContent: string | null = null;
+    databaseSchema: string;
+    outStream?: Writable;
+    continuityAgent?: ContinuityAgent;
+    synopsisAgent?: SynopsisAgent;
+    lastUserMessage: string;
+    max_message_pairs: number;
+    max_csum_tokens: number;
+    max_similarity_tokens: number;
+    max_synopses_tokens: number;
+    asyncResults: Promise<FunctionCallResult[]>;
+    continuityResponsePromise: Promise<string> | null = null;
+    continuityResponseContent: string | null = null;
 
     constructor(args: MementoAgentArgs)
     {
@@ -102,27 +94,6 @@ export class MementoAgent extends Agent
         return assistantMessage;
     }
 
-    async retrieveContext(): Promise<MementoPromptTemplateArgs> {
-        const pinnedCsumMems: Memento[] = await this.DB.searchPinnedCsumMems(this.max_csum_tokens);
-        const synopses: string[] = await this.DB.getSynopses(this.max_synopses_tokens);
-        const selectedMems: SimilarityResult[] = await this.DB.searchMemsBySimilarity(this.lastUserMessage, this.max_similarity_tokens);
-        const totalTokens: number = selectedMems.reduce((acc, mem) => acc + mem.tokens, 0);
-        const continuityResponseContent: string | null = this.continuityResponseContent;
-        const functions = functionCallingInstructions(this.databaseSchema);
-
-        mlog(`selectedMems results: length: ${selectedMems.length}, total tokens: ${totalTokens}`);
-
-        return {
-            system: mementoCoreSystemPrompt,
-            functions,
-            databaseSchema: this.databaseSchema,
-            pinnedCsumMems,
-            synopses,
-            selectedMems,
-            continuityResponseContent
-        };
-    }
-
     checkForFunctionResults(content: string): void {
         if (!content.startsWith(FUNCTION_RESULT_HEADER)) {
             this.lastUserMessage = content;
@@ -140,7 +111,7 @@ export class MementoAgent extends Agent
 
         await this.DB.addConversationMem(newMessage);
 
-        const tempateArgs: MementoPromptTemplateArgs = await this.retrieveContext();
+        const tempateArgs: MementoPromptTemplateArgs = await retrieveContext(this);
         const prompt = mementoPromptTemplate(tempateArgs);
 
         const messages: Message[] = [...priorConveration, newMessage];
